Issue 262/263 - Multiple Passed/Failed Statements
Per - cmi5 Working Group Meeting Minutes - October 9th, 2015

Multiple Scored/Passed/Complete/Fail Statements

The group agreed that there will be only one (cmi5-defined) statement allowed in an AU session for Completed. And only one (cmi5-defined) statement allowed in an AU session for Passed OR Failed.

As a result of this decision the following changes were agreed to:

Update section 9.3 to add the following language:

The following xAPI verb requirements in this section are for “cmi5 defined” statements in this specification.
These verbs MAY also be used for “cmi5 allowed” statements per section 9.6.2.
Note that “cmi5 allowed” statements are not subject to the usage rules in this section.
Replace “statement” with “cmi5 defined statement” – in the verb definitions of section 9.3.

Remove “PassIsFinal” (all sections and XSD)

Update sections 9.3.4 Passed and 9.3.5 Failed to Limit Failed or Passed to one per session. (Cannot have both Passed and Failed in same session for “cmi5 defined” statements)

--------

Note:
1.  additional language added to refer to "any statements". (cmi5 defined or cmi5 allowed)
2. "passIsFinal" not removed from all examples - followup pull requested needed