Update cmi5_spec with changes to PR #2 from Nov 18, 2022 meeting.md
Added revisions decided on at Nov 18, 2022 cmi5 Working group meeting.

Pull Request #2 - The group agreed to the following changes:

9.3.7 Waived - Capitalize MAY
>>DONE

10.2.3 launchParameters - Capitalize MAY
>>DONE

13.1.4 AU Metadata - Capitalize MAY in section and in launchMethod
>>DONE

5.0 Conceptual Model: Informative

An LMS imports a course structure which contains at least one AU. Optionally, the course structure can include one or more blocks, which consist of 1 or more AUs or nested blocks.
>>DONE

AUs MAY use additional verbs not listed in this specification.
>>DONE

9.5.1 Score

A score MAY be reported. If a score is reported by an AU, the verb MUST be consistent with "masteryScore" (if defined for the AU in the LMS Launch Data).
Capitalization and quoting of concepts is not done consistently throughout the cmi5 Profile Specification document #773
>>DONE