issue #803 revise references to standard back to specification as per Aug 25, 2023 meeting
sorry about the long commit description but I want to make sure I have done everything correctly so I documented it as I went along.

This commit has changes that were decided at the Aug 25, 2023 meeting.

There were discussions at both the Aug 18 and Aug 25, 2023 sub group meetings about how to refer to xAPI in the cmi5 specification I.e. should there be a reference to xAPI 2.0, should it be referred to as a standard or a specification.

At the Aug 18, 2023 meeting it was decided to add the a reference to xAPI 2.0 and update the definition of Experience API (xAPI) to refer to xAPI 2.0.  The new text was added.

References
     * Experience API (xAPI) Standard (IEEE 9274.1.1 xAPI 2.0)
     https://opensource.ieee.org/xapi

Definitions
* __Experience API (xAPI)__: The IEEE 9274.1.1 xAPI 2.0 Standard. A runtime data communication specification for learning content (AU) to send and receive data to a Learning Record Store (LRS). xAPI is used to define the data transport and the data model. As far as requirements found in this document (other than the requirement of xAPI usage), the xAPI Specification 1.0.x is equivalent.

At the Aug 25, 2023 meeting there was further discussion on the topic of  whether to refer to the new xAPI 2.0 standard in the cmi5 Specification and it was decided that it was premature and so the decision was made to reverse the changes made at the Aug 18 meeting.

References
- The new reference to xAPI 2.0 was removed.
- The URL for xAPI in the original reference was changed to
     https://github.com/adlnet/xAPI-Spec

Definitions
The references to xAPI 2.0  and xAPI Specification 1.0.x were removed from the Definition of Experience API (xAPI). Also the second sentence was edited for clarity.  The revised text is as follows:

* __Experience API (xAPI)__: A runtime data communication specification for learning content (AU) to send and receive data to a Learning Record Store (LRS).  xAPI is used to define the data transport and the data model.

Other decisions at the Aug 25, 2023 meeting related to the discussions above.

- Change “xAPI specification” references to just “xAPI” (not “xAPI standard”)s)
- Change “xAPI specification” references (outside of 2.0 References and 3.0 Definitions to just “xAPI”
- Continue to refer to cmi5 as a “specification” on the current Github site.
	- there were a few references to “this standard” referring to the cmi5 specification that I changed to ‘this specification’.