Update cmi5_spec.md re issue #806
Issue #806   related changes

**Full Section names and numbers**

I made changes to any references to sections following the format described above with the full section number and then the section name.
- Any references to Section 10 I edited replacing them with the wording above"
- There were many instances the Section was referenced by only the word "Section" and section number.
     - for all of these I added the section name.

**Em dashes in Section references**

In the following instances the reference to other sections had an em dash '–' between the section number and the section name:

>4.1 Assignable Unit (AU)
>
> See Section 7.0 – AU Requirements. ...

>4.2 Learning Management Systems
>
> See Section 6.0 – LMS Requirements. ...

>10.2.5 moveOn
>
>...
>Value Space:</th><td><code>moveOn</code> values as defined in the Course Structure (Section 13.1.4 – AU Metadata)

To maintain consistency with the other section references I removed the em dashes from those instances above.

**Em dashes in subsection references**

In another instance I noticed an em dash in a section reference  but it was after the section name and number but before a subsection name, in this case the subsection was the ID field in the table.

9.6.2.3 Publisher ID Grouping Activity

> ...
> ((See Section 13.1.4 AU Metadata – ID)...
> ...

refers to

>13.1.4 AU Metadata
>>id

So in this case I left the em dash in the section  reference.