Issue #806 Update cmi5_spec.md - Section 9.3.1 Launched  - lms oblications
Issue #806

Section 9.3.1 Launched > LMS Obligations.

There was one instance where I wasn't sure how to revise it.
See Section 9.3.1 Launched > LMS Obligations.
(See Statement API, Section 10)

For now I just changed the reference to Section 10 to the full Section name and number.

      - "See Statement API, Section 10.0 xAPI State Data Model"

but it refers also to "Statement API".  is that another section or is just part of the reference to Section 10?

Is that correct?